Assistant checkpoint
Assistant generated file changes:
- models.py: Add profile fields to User model
- main.py: Add profile routes
- templates/profile.html: Create profile template
- templates/home.html: Add profile menu

---

User prompt:

ok now i want there to be a profile icon in the top right corner when the user is signed in at all times (the user can click on this and there will be two options: profile settings and logout). For the profile settings, the user can choose a profile picture (either one of the avliale school logos, a picture of a water polo ball, or any picture they choose), they can choose to follow a certain team (s) (which will put a start next to that team on the team page and will put that team's upcoming games on the home page).